Restore static binders: they must stay in org.sfl4j.impl
