+++ src/test/java/org/slf4j/impl/StaticMDCBinderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.slf4j.simple;
+package org.slf4j.impl;
 
 import org.junit.jupiter.api.Test;
